refactor: share GitHub auth redirect URL helpers
