+++ app/src/ai/ambient_agents/mod.rs
@@ -8,6 +8,7 @@ use std::str::FromStr;
 use uuid::{NonNilUuid, Uuid};
 
 pub mod github_auth_notifier;
+pub mod github_auth_url;
 pub mod scheduled;
 pub mod spawn;
 pub mod task;
