+++ app/src/settings_view/update_environment_form.rs
@@ -4,7 +4,10 @@ use super::{
 };
 use crate::server::server_api::ServerApiProvider;
 use crate::{
-    ai::ambient_agents::telemetry::CloudAgentTelemetryEvent,
+    ai::ambient_agents::{
+        github_auth_url::{self, AuthSource, GithubAuthRedirectTarget},
+        telemetry::CloudAgentTelemetryEvent,
+    },
     ai::{
         ambient_agents::github_auth_notifier::{GitHubAuthEvent, GitHubAuthNotifier},
         cloud_environments::{AmbientAgentEnvironment, GithubRepo},
@@ -149,21 +152,6 @@ pub enum EnvironmentFormMode {
     Edit { env_id: SyncId },
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum GithubAuthRedirectTarget {
-    SettingsEnvironments,
-    FocusCloudMode,
-}
-
-impl GithubAuthRedirectTarget {
-    fn next_path(self) -> &'static str {
-        match self {
-            Self::SettingsEnvironments => "settings/environments",
-            Self::FocusCloudMode => "action/focus_cloud_mode",
-        }
-    }
-}
-
 /// Events emitted by UpdateEnvironmentForm.
 #[derive(Debug, Clone)]
 pub enum UpdateEnvironmentFormEvent {
@@ -255,23 +243,6 @@ enum SuggestImageState {
         key: String,
         message: String,
     },
-}
-
-/// Indicates where the GitHub authorization flow was initiated from.
-/// This affects the redirect URL used after auth completes.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-pub enum AuthSource {
-    /// Auth initiated from the settings page (default behavior: redirect to settings)
-    #[default]
-    Settings,
-    /// Auth initiated from cloud agent setup (skip redirect, just refresh in place)
-    CloudSetup,
-}
-
-#[derive(Clone, Copy, Debug)]
-enum OAuthNextPlatform {
-    Native,
-    Web,
 }
 
 pub struct UpdateEnvironmentForm {
@@ -2715,13 +2686,17 @@ impl UpdateEnvironmentForm {
     }
 
     fn auth_url_with_next(&self, base_auth_url: &str) -> String {
-        let scheme = Self::oauth_next_scheme();
-        Self::build_auth_url_with_next_internal(
-            base_auth_url,
-            self.github_auth_redirect_target,
-            &scheme,
-            self.auth_source,
-        )
+        match (self.github_auth_redirect_target, self.auth_source) {
+            (GithubAuthRedirectTarget::SettingsEnvironments, AuthSource::Settings) => {
+                github_auth_url::settings_environments_auth_url_with_next(base_auth_url)
+            }
+            (GithubAuthRedirectTarget::FocusCloudMode, AuthSource::CloudSetup) => {
+                github_auth_url::cloud_setup_auth_url_with_next(base_auth_url)
+            }
+            (target, auth_source) => {
+                github_auth_url::auth_url_with_next(base_auth_url, target, auth_source)
+            }
+        }
     }
 
     #[cfg(test)]
@@ -2730,107 +2705,12 @@ impl UpdateEnvironmentForm {
         target: GithubAuthRedirectTarget,
         scheme: &str,
     ) -> String {
-        Self::build_auth_url_with_next_internal(base_auth_url, target, scheme, AuthSource::Settings)
-    }
-
-    fn build_auth_url_with_next_internal(
-        base_auth_url: &str,
-        target: GithubAuthRedirectTarget,
-        scheme: &str,
-        auth_source: AuthSource,
-    ) -> String {
-        let Ok(mut url) = Url::parse(base_auth_url) else {
-            return base_auth_url.to_string();
-        };
-
-        let scheme_for_next = std::env::var("WARP_OAUTH_NEXT_SCHEME")
-            .ok()
-            .filter(|value| !value.is_empty())
-            .or_else(|| {
-                url.query_pairs()
-                    .find(|(key, _)| key == "scheme")
-                    .map(|(_, value)| value.into_owned())
-            })
-            .filter(|value| !value.is_empty())
-            .unwrap_or_else(|| scheme.to_string());
-
-        let platform = if cfg!(target_family = "wasm") {
-            OAuthNextPlatform::Web
-        } else {
-            OAuthNextPlatform::Native
-        };
-
-        let next_url = Self::build_next_url(target, &scheme_for_next, auth_source, platform)
-            .unwrap_or_else(|| format!("{scheme_for_next}://{}", target.next_path()));
-
-        let existing_pairs = url
-            .query_pairs()
-            .filter(|(key, _)| key != "next")
-            .map(|(key, value)| (key.into_owned(), value.into_owned()))
-            .collect::<Vec<_>>();
-
-        {
-            let mut query_pairs = url.query_pairs_mut();
-            query_pairs.clear();
-            for (key, value) in existing_pairs {
-                query_pairs.append_pair(&key, &value);
-            }
-            query_pairs.append_pair("next", &next_url);
-        }
-
-        url.to_string()
-    }
-
-    fn build_next_url(
-        target: GithubAuthRedirectTarget,
-        scheme_for_next: &str,
-        auth_source: AuthSource,
-        platform: OAuthNextPlatform,
-    ) -> Option<String> {
-        match platform {
-            OAuthNextPlatform::Native => {
-                let base = format!("{scheme_for_next}://{}", target.next_path());
-                let mut url = Url::parse(&base).ok()?;
-
-                if matches!(auth_source, AuthSource::CloudSetup) {
-                    url.query_pairs_mut()
-                        .append_pair("source", crate::uri::CLOUD_SETUP_SOURCE);
-                }
-
-                Some(url.to_string())
-            }
-            OAuthNextPlatform::Web => {
-                let mut url = Url::parse(&ChannelState::server_root_url()).ok()?;
-                url.set_query(None);
-
-                match target {
-                    GithubAuthRedirectTarget::SettingsEnvironments => {
-                        url.set_path("/settings/environments");
-                        {
-                            let mut pairs = url.query_pairs_mut();
-                            pairs.append_pair("oauth", "github");
-                            if matches!(auth_source, AuthSource::CloudSetup) {
-                                pairs.append_pair("source", crate::uri::CLOUD_SETUP_SOURCE);
-                            }
-                        }
-                    }
-                    GithubAuthRedirectTarget::FocusCloudMode => {
-                        url.set_path("/action/focus_cloud_mode");
-                    }
-                }
-
-                Some(url.to_string())
-            }
-        }
-    }
-
-    fn oauth_next_scheme() -> String {
-        if let Ok(override_value) = std::env::var("WARP_OAUTH_NEXT_SCHEME") {
-            if !override_value.is_empty() {
-                return override_value;
-            }
-        }
-        ChannelState::url_scheme().to_string()
+        github_auth_url::build_auth_url_with_next(
+            base_auth_url,
+            target,
+            scheme,
+            AuthSource::Settings,
+        )
     }
 
     /// Parses a Docker image reference and returns the Docker Hub URL if it looks like a Docker Hub image.
